chore: move all ts code to src folder
