+++ dev-client/src/components/common/SearchBar.tsx
